refactor: inline mixin logic into vaadin-rich-text-editor-popup
